backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.0
django-cleanup==8.0.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
drf-spectacular==0.26.5
factory-boy==3.3.0
Faker==19.13.0
inflection==0.5.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.12.0
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
